requirements.txt
# NOTE: for GPU experiments (StackOverflow / MovieLens), please install `jax[cuda]`;
# see https://github.com/google/jax#pip-installation-gpu-cuda for instructions.
jax[cpu]==0.3.20
dm-haiku==0.0.8
optax==0.1.3
matplotlib
numpy
pandas
scipy
tensorflow_privacy==0.7.3
tensorflow  # TF is needed mainly for data utils
tensorflow_datasets
tqdm